requirements.txt
# ==== requirements.txt (re-ordered & compatible) ====
numpy==2.3.0          
pandas==2.3.1      
pyarrow==20.0.0       

# Core data processing
requests==2.31.0
python-dotenv==1.0.0

# New FMP ingestion requirements
pyspark==3.5.0
delta-spark==3.0.0
aiohttp==3.8.6
great-expectations==0.18.8
pydantic==2.7.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Existing Django app deps
asgiref==3.8.1
Django==5.2
fastparquet==2024.11.0
fsspec==2025.5.1
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.10
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
cramjam==2.10.0
